+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 % erl-cmdline changelog
 
 # Next Version
+## Misc
+- Introduce `cmdline:config/0` as alias for `cmdline_config:config/0`.
 
 # 1.0.0
 First public version.